More movement, reclaim some validation we didn't get via ingot.
