+++ lib/opte-test-utils/src/lib.rs
@@ -33,11 +33,11 @@ pub use opte::engine::ingot_packet::MblkLiteParsed;
 pub use opte::engine::ingot_packet::MsgBlk;
 pub use opte::engine::ingot_packet::Packet2;
 pub use opte::engine::ip::v4::Ipv4;
+pub use opte::engine::ip::v4::Ipv4Addr;
+pub use opte::engine::ip::v4::Protocol;
 pub use opte::engine::ip::v6::Ipv6;
+pub use opte::engine::ip::v6::Ipv6Addr;
 pub use opte::engine::ip::L3Repr;
-pub use opte::engine::ip4::Ipv4Addr;
-pub use opte::engine::ip4::Protocol;
-pub use opte::engine::ip6::Ipv6Addr;
 pub use opte::engine::layer::DenyReason;
 pub use opte::engine::packet::ParseError;
 pub use opte::engine::port::meta::ActionMeta;
