+++ lib/opte/src/engine/predicate.rs
@@ -15,14 +15,14 @@ use super::icmp::v6::MessageType as Icmpv6MessageType;
 use super::ingot_packet::ulp_dst_port;
 use super::ingot_packet::ulp_src_port;
 use super::ingot_packet::MblkPacketData;
+use super::ip::v4::Ipv4Addr;
+use super::ip::v4::Ipv4Cidr;
 use super::ip::v4::Ipv4Ref;
+use super::ip::v4::Protocol;
+use super::ip::v6::Ipv6Addr;
+use super::ip::v6::Ipv6Cidr;
 use super::ip::v6::Ipv6Ref;
 use super::ip::L3;
-use super::ip4::Ipv4Addr;
-use super::ip4::Ipv4Cidr;
-use super::ip4::Protocol;
-use super::ip6::Ipv6Addr;
-use super::ip6::Ipv6Cidr;
 use super::port::meta::ActionMeta;
 use alloc::boxed::Box;
 use alloc::string::String;
