+++ lib/oxide-vpc/src/engine/overlay.rs
@@ -43,10 +43,10 @@ use opte::engine::headers::IpAddr;
 use opte::engine::headers::IpCidr;
 use opte::engine::headers::IpPush;
 use opte::engine::ingot_packet::MblkPacketData;
-use opte::engine::ip4::Protocol;
-use opte::engine::ip6::Ipv6Addr;
-use opte::engine::ip6::Ipv6Cidr;
-use opte::engine::ip6::Ipv6Push;
+use opte::engine::ip::v4::Protocol;
+use opte::engine::ip::v6::Ipv6Addr;
+use opte::engine::ip::v6::Ipv6Cidr;
+use opte::engine::ip::v6::Ipv6Push;
 use opte::engine::layer::DefaultAction;
 use opte::engine::layer::Layer;
 use opte::engine::layer::LayerActions;
