+++ lib/oxide-vpc/tests/integration_tests.rs
@@ -30,12 +30,12 @@ use opte::engine::geneve::Vni;
 use opte::engine::ingot_packet::MblkFullParsed;
 use opte::engine::ingot_packet::MsgBlk;
 use opte::engine::ingot_packet::Packet2;
+use opte::engine::ip::v4::Ipv4Addr;
 use opte::engine::ip::v4::Ipv4Ref;
 use opte::engine::ip::v6::Ipv6;
 use opte::engine::ip::v6::Ipv6Ref;
 use opte::engine::ip::ValidL3;
 use opte::engine::ip::L3;
-use opte::engine::ip4::Ipv4Addr;
 use opte::engine::packet::InnerFlowId;
 use opte::engine::parse::ValidUlp;
 use opte::engine::port::ProcessError;
